Common setup for Play apps
