+++ apps/rule-manager/app/AppComponents.scala
@@ -27,7 +27,12 @@ class AppComponents(context: Context, region: String, identity: AppIdentity, cre
   val config = new RuleManagerConfig(configuration, region, identity, creds)
   val db = new RuleManagerDB(config.dbUrl, config.dbUsername, config.dbPassword)
 
-  private val s3Client = AmazonS3ClientBuilder.standard().withCredentials(creds).build()
+  private val s3Client = AmazonS3ClientBuilder
+    .standard()
+    .withCredentials(creds)
+    .withRegion(region)
+    .build()
+
   val stageDomain = identity match {
     case identity: AwsIdentity if identity.stage == "PROD" => "gutools.co.uk"
     case identity: AwsIdentity => s"${identity.stage.toLowerCase}.dev-gutools.co.uk"
